Create section3 .10
This code creates an array of integers and then calls the 'reverseArray' method to reverse the array. The 'reverseArray' method creates a new array and assigns the elements of the original array to it in reverse order using a loop. The reversed array is then returned and printed to the console.

The output of this code will be: 5 4 3 2 1